requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.33.0
pydantic-settings==2.6.0
sqlalchemy==2.0.36
alembic==1.13.3
psycopg[binary]==3.2.2
redis==5.1.1
celery==5.4.0
pyjwt[crypto]==2.9.0
bcrypt==4.2.0
python-multipart==0.0.17
httpx==0.27.2
pgvector==0.2.5
email-validator==2.2.0
